requirements.txt
# Core dependencies
pyyaml>=6.0.1

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Network analysis (for graph visualization)
networkx>=3.0

# Optional: For advanced URL parsing
urllib3>=2.0.0

# Development dependencies
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Documentation
markdown>=3.4.0
